server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfplumber>=0.10.3
PyMuPDF>=1.23.8
google-cloud-texttospeech>=2.14.2
google-cloud-storage>=2.14.0
moviepy<2.0.0
Pillow>=10.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
google-generativeai>=0.3.0

